requirements/local.txt
# Local development dependencies go here
-r base.txt
django-debug-toolbar==2.2
django-extensions==2.2.7
